gate/requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
environs==9.3.1
fastapi==0.70.0
h11==0.12.0
idna==3.1
marshmallow==3.11.1
multidict==5.1.0
orjson==3.5.0
pre-commit==2.12.0
pycparser==2.20
PyJWT==2.0.1
python-dotenv==0.17.0
starlette==0.16.0
typing-extensions==3.7.4.3
urllib3==1.26.4
uvicorn==0.13.4
yarl==1.6.3
python-json-logger==2.0.1
